myproject/requirements.txt
Django==5.0.6
djangorestframework_simplejwt==5.3.1
python-decouple==3.8
djangorestframework==3.15.2
